Fix bug preventing encoding pointer to time.Time
The encoder handles the case where a field is of type
time.Time. Additionally, non-nil pointer values are encoded as the
value pointed to by drilling down to the element the pointer field
references. However, the ordering of which this was done was
preventing the combination of the two from working.